backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
openai>=1.0,<2.0
chromadb>=0.5,<1.0
python-dotenv>=1.0,<2.0
pydantic>=2.0,<3.0
pdfplumber>=0.11,<1.0
